requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.7
Werkzeug==2.3.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-dotenv==1.0.1